Conflicts:
	.env
	.gitignore

+++ src/services/mobileExperienceService.js
@@ -1,31 +1,46 @@
 const Experience = require('../models/Experience')
 const FileExperienceMapping = require('../models/FileExperienceMapping')
 const ServiceHelper = require('../helper/ServiceHelper')
-const fileService = require('../services/fileService')
 const { raw } = require('objection');
+const { UnsupportedOperationError, NotFoundError } = require('../models/errors')
+const fileService = require('../services/fileService')
 
 const experienceService = {}
 
-experienceService.createExperience = async (experienceDTO, loggedInUser, fileIds) => {
+const UnsupportedOperationErrorEnum = {
+    FILE_NOT_FOUND: 'FILE_NOT_FOUND'
+}
 
+experienceService.createExperience = async (experienceDTO, loggedInUser, fileId) => {
+
+    let file 
+    if( fileId !== 0 ) {
+        
+        file = await fileService.getFileByIdAndCreateBy(fileId, loggedInUser.sub)
+
+        if(!file)
+            throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.FILE_NOT_FOUND)
+
+    }
+  
+    // insert file to mapping
     const experience = await Experience.query()
     .insertToTable(experienceDTO, loggedInUser.sub)
 
-    if( fileIds !== undefined ) {
-        const mapping = fileIds.map(fileId => ({
+    // if file exist
+    if(file) {
+        await FileExperienceMapping.query()
+        .insertToTable({
             efileefileid: fileId,
             eexperienceeexperienceid: experience.eexperienceid
-        }))
-        
-        // insert file to mapping
-        await FileExperienceMapping.query().insertToTable(mapping, loggedInUser.sub)
+        }, loggedInUser.sub)
     }
 
     return experience
 
 }
 
-experienceService.getExperienceList = async (page =0, size = 10, loggedInUser, keyword) => {
+experienceService.getExperienceList = async (page = 0, size = 10, loggedInUser, keyword) => {
 
     let newKeyword = ''
 
@@ -41,54 +56,68 @@ experienceService.getExperienceList = async (page =0, size = 10, loggedInUser, k
     return ServiceHelper.toPageObj(page, size, pageObj)
 }
 
-experienceService.getExperienceById = async (experienceId, loggedInUser) => {
+experienceService.getExperienceById = async (experienceId) => {
 
-    const experience = await Experience.query()
-    .select('eexperienceid','eexperiencename', 'eexperiencestartdate', 'eexperienceenddate', 'eexperiencelocation', 'eexperienceposition', 'eexperiencedescription', 'eindustryname')
-    .joinRelated('industries')
+    return Experience.query()
+    .select('eexperienceid','eexperiencename', 'eexperiencestartdate', 'eexperienceenddate', 'eexperiencelocation', 
+    'eexperienceposition', 'eexperiencedescription', 'eindustryname', 'efileid', 'efilename')
+    .leftJoinRelated('[industries, files]')
     .where('eexperienceid', experienceId)
-    .where('eusereuserid', loggedInUser.sub)
     .first()
-
-    const files = await FileExperienceMapping.query()
-    .select('efileefileid')
-    .where('eexperienceeexperienceid', experienceId)
-    .where('efileexperiencemappingcreateby', loggedInUser.sub)
-
-    const result = {
-        ...experience,
-        files
-    }
-
-    return result
 
 }
 
-experienceService.editExperience = async (experienceDTO, experienceId, loggedInUser, fileIds) => {
+experienceService.editExperience = async (experienceDTO, experienceId, loggedInUser, fileId) => {
 
-    // remove file experience mapping by experienceid
+    const experience = await experienceService.getExperienceById(experienceId);
+
+    if (!experience)
+        throw new NotFoundError()
+
+    let promises = []
+
+    // remove previous file, so when edit file to 'no file', file already deleted
     await FileExperienceMapping.query()
     .delete()
     .where('eexperienceeexperienceid', experienceId)
 
-    if( fileIds !== undefined ) {
-        const mapping = fileIds.map(fileId => ({
-            efileefileid: fileId,
-            eexperienceeexperienceid: experienceId
-        }))
-    
-        // add file experience mapping
-        await FileExperienceMapping.query()
-        .insertToTable(mapping, loggedInUser.sub)
+    let file 
+
+    if( fileId !== 0 ) {
+        
+        file = await fileService.getFileByIdAndCreateBy(fileId, loggedInUser.sub)
+
+        if(!file)
+            throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.FILE_NOT_FOUND)
+
     }
 
+    if(file) {
+        promises.push( 
+            FileExperienceMapping.query()
+            .insertToTable({
+                efileefileid: fileId,
+                eexperienceeexperienceid: experienceId
+            }, loggedInUser.sub)
+        ) 
+    }
 
-    return Experience.query()
-    .where('eexperienceid', experienceId)
-    .where('eusereuserid', loggedInUser.sub)
-    .first()
-    .updateByUserId(experienceDTO, loggedInUser.sub)
-    .returning('*')
+    promises.push(
+        Experience.query()
+        .where('eexperienceid', experienceId)
+        .where('eusereuserid', loggedInUser.sub)
+        .first()
+        .updateByUserId(experienceDTO, loggedInUser.sub)
+        .returning('*')
+    )
+
+    return Promise.all(promises)
+    .then(arr => {
+        // if promises length = 1, it means no FileExperienceMapping promise pushed
+        if(promises.length === 1)
+            return arr[0]
+        return arr[1]
+    })
 
 }
 
@@ -99,6 +128,7 @@ experienceService.deleteExperience = async (experienceId, loggedInUser) => {
     .where('eexperienceid', experienceId)
     .where('eusereuserid', loggedInUser.sub)
     .then(rowsAffected => rowsAffected === 1)
+
 }
 
 module.exports = experienceService
