+++ src/services/mobileUserService.js
@@ -1,13 +1,36 @@
 require('dotenv').config();
 const User = require('../models/User');
 const CoachIndustryMapping = require('../models/CoachIndustryMapping')
+const UserIndustryMapping = require('../models/UserIndustryMapping')
 const bcrypt = require('../helper/bcrypt');
 const jwt = require('jsonwebtoken');
 const fileService = require('./fileService');
 const Otp = require('../models/Otp');
 const emailService = require('../helper/emailService');
+const { UnsupportedOperationError, NotFoundError } = require('../models/errors')
+const CompanyLog = require('../models/CompanyLog')
+const CompanyUserMapping = require('../models/CompanyUserMapping');
+const TeamUserMapping = require('../models/TeamUserMapping');
+const License = require('../models/License');
+const ServiceHelper = require('../helper/ServiceHelper')
+const CompanyLogTypeEnum = require('../models/enum/CompanyLogTypeEnum')
+const CompanyLogStatusEnum = require('../models/enum/CompanyLogStatusEnum')
 
 const UserService = {};
+
+const UnsupportedOperationErrorEnum = {
+    NOT_ADMIN: 'NOT_ADMIN',
+    USER_NOT_EXIST: 'USER_NOT_EXIST',
+    INDUSTRY_IS_EMPTY: 'INDUSTRY_IS_EMPTY',
+}
+
+const ErrorUserEnum = {
+    EMAIL_NOT_FOUND :'EMAIL_NOT_FOUND',
+    EMAIL_INVALID : 'EMAIL_INVALID',
+    USER_ALREADY_EXIST : 'USER_ALREADY_EXIST',
+    OTP_NOT_FOUND : 'OTP_NOT_FOUND',
+    OTP_CODE_NOT_MATCH : 'OTP_CODE_NOT_MATCH'
+}
 
 async function generateJWTToken(user) {
 
@@ -28,13 +51,11 @@ UserService.login = async (loginDTO) => {
 
     const user = await User.query().where('euseremail', loginDTO.euseremail).first();
 
-    if (!user)
-        return
+    if (!user) throw new UnsupportedOperationError(ErrorUserEnum.EMAIL_NOT_FOUND)
 
     const success = await bcrypt.compare(loginDTO.euserpassword, user.euserpassword);
 
-    if (!success)
-        return
+    if (!success) throw new UnsupportedOperationError('SERVER_ERROR')
 
     return await generateJWTToken(user);
 
@@ -45,20 +66,20 @@ UserService.createUser = async (userDTO, otpCode) => {
     const isEmail = emailService.validateEmail(userDTO.euseremail);
 
     if (!isEmail)
-        return 'invalid email'
+        throw new UnsupportedOperationError(ErrorUserEnum.EMAIL_INVALID)
 
     const user = await User.query().where('euseremail', userDTO.euseremail).first();
 
     if (user)
-        return 'user exist'
+        throw new UnsupportedOperationError(ErrorUserEnum.USER_ALREADY_EXIST)
 
     const otp = await Otp.query().where('euseremail', userDTO.euseremail).first();
 
     if (!otp)
-        return 'no otp found'
+        throw new UnsupportedOperationError(ErrorUserEnum.OTP_NOT_FOUND)
 
     if (otp.eotpcode !== otpCode)
-        return 'otp code not match'
+        throw new UnsupportedOperationError(ErrorUserEnum.OTP_CODE_NOT_MATCH)
 
     // confirm OTP
     await otp.$query().updateByUserId({ eotpconfirmed: true }, 0);
@@ -75,13 +96,35 @@ UserService.getUserById = async (userId) => {
     .select('euserid', 'eusername', 'eusermobilenumber', 'euseremail', 'euseridentitynumber', 'euserdob', 'euseraddress', 'eusergender', 
     'euserhobby', 'euserfacebook', 'euserinstagram', 'euserlinkedin', 'ecountryname', 'efileefileid', 'euseriscoach')
     .leftJoinRelated('country')
-    .where('euserid', userId).first();
+    .where('euserid', userId)
+    .first();
 
     if (!user)
         return
 
     return user;
     
+}
+
+UserService.getOtherUserById = async (userId, type) => {
+
+    if (type !== 'USER' && type !== 'COACH')
+        return
+
+    let relatedIndustry = 'coachIndustries'
+    if (type === 'USER')
+        relatedIndustry = 'userIndustries'
+
+    return User.query()
+    .findById(userId)
+    .modify('baseAttributes')
+    .withGraphFetched(relatedIndustry)
+    .withGraphFetched('companies(baseAttributes)')
+    .withGraphFetched('teams(baseAttributes)')
+    .withGraphFetched('experiences(baseAttributes)')
+    .withGraphFetched('licenses(baseAttributes)')
+    // .withGraphFetched("[" + relatedIndustry + ", companies(baseAttributes), teams(baseAttributes), experiences.industry, licenses.industry]")
+
 }
 
 async function updateUserAndIndustries(userFromDB, userDTO, industryIds, user, trx) {
@@ -94,6 +137,8 @@ async function updateUserAndIndustries(userFromDB, userDTO, industryIds, user, t
             eindustryeindustryid: industryId
         }
     });
+
+    await UserIndustryMapping.query().where('eusereuserid', user.sub).delete();
 
     return userFromDB.$relatedQuery('userIndustriesMapping', trx).insertToTable(userIndustryMapping, user.sub);
 
@@ -117,9 +162,15 @@ UserService.updateUser = async (userDTO, industryIds, user) => {
     if (!userFromDB)
         return
     
-    await User.transaction(async trx => {
-        await updateUserAndIndustries(userFromDB, userDTO, industryIds, user, trx);
-    })
+    // Update user only
+    if (industryIds.length === 0) {
+        await userFromDB.$query().updateByUserId(userDTO, user.sub);
+    } else {
+        // Update user and industry
+        await User.transaction(async trx => {
+            await updateUserAndIndustries(userFromDB, userDTO, industryIds, user, trx);
+        })
+    }
 
     return 1;
 
@@ -131,8 +182,9 @@ UserService.updateUserCoachData = async (userCoachDTO, user, industryIds) => {
 
     if (!userFromDB)
         return
-    
-    userCoachDTO.euserdob = new Date(userCoachDTO.euserdob).getTime();
+
+    if (userFromDB.euseriscoach)
+        return
 
     const coachIndustryMappings = industryIds.map(industryId => ({
         eindustryeindustryid: industryId,
@@ -154,7 +206,10 @@ UserService.removeCoach = async (user) => {
     const userFromDB = await UserService.getUserById(user.sub);
 
     if (!userFromDB)
-        return 
+        return
+
+    if (userFromDB.euseriscoach === false)
+        return
 
     await CoachIndustryMapping.query()
     .delete()
@@ -180,6 +235,91 @@ UserService.changePassword = async (oldPassword, newPassword, user) => {
     const hashedNewPassword = await bcrypt.hash(newPassword);
 
     return userFromDB.$query().updateByUserId({euserpassword: hashedNewPassword}, user.sub);
+
+}
+
+UserService.getIndustryByUserId = async (user, type) => {
+
+    const userFromDB = await UserService.getUserById(user.sub)
+
+    if(!userFromDB)
+        throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.USER_NOT_EXIST)
+
+    if(type === 'USER') {
+
+        return UserIndustryMapping.query()
+            .select('eindustryid', 'eindustryname')
+            .joinRelated('industry')
+            .where('eusereuserid', user.sub);
+
+    } else if ( type === 'COACH') {
+
+        return CoachIndustryMapping.query()
+            .select('eindustryid', 'eindustryname')
+            .joinRelated('industry')
+            .where('eusereuserid', user.sub);
+    }
+
+}
+
+UserService.changeIndustryByUserId = async (user, type, industryIds) => {
+
+    const userFromDB = await UserService.getUserById(user.sub)
+
+    if(!userFromDB)
+        throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.USER_NOT_EXIST)
+
+    if(industryIds.length <= 0) 
+        throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.INDUSTRY_IS_EMPTY)
+
+    const mapping = industryIds.map(industryId => ({
+        eusereuserid: user.sub,
+        eindustryeindustryid: industryId
+    }))
+
+    if(type === 'USER') {
+
+        return UserIndustryMapping.query()
+            .delete()
+            .where('eusereuserid', user.sub)
+            .then(ignore => {
+                    return UserIndustryMapping.query()
+                    .insertToTable(mapping, user.sub)
+                })
+
+    } else if ( type === 'COACH') {
+
+        return CoachIndustryMapping.query()
+            .delete()
+            .where('eusereuserid', user.sub)
+            .then(ignore => {
+                    return CoachIndustryMapping.query()
+                    .insertToTable(mapping, user.sub)
+                })
+    }
+
+}
+
+UserService.getListPendingByUserId = async (page, size, userId, type) => {
+
+    const userFromDB = User.query()
+        .select()
+        .where('euserid', userId)
+        .first()
+    
+    if(!userFromDB)
+        throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.USER_NOT_EXIST)
+
+    if( type !== CompanyLogTypeEnum.INVITE && type !== CompanyLogTypeEnum.APPLY)
+        throw new NotFoundError()
+
+    return CompanyLog.query()
+    .where('eusereuserid', userId)
+    .where('ecompanylogtype', type)
+    .andWhere('ecompanylogstatus', CompanyLogStatusEnum.PENDING)
+    .orderBy('ecompanylogcreatetime', 'DESC')
+    .page(page, size)
+    .then(pageObj => ServiceHelper.toPageObj(page, size, pageObj))
 
 }
